feat(DashboardPresentation): modularize
