+++ src/components/dashboard/components/index.ts
@@ -1,0 +1,3 @@
+export { default as WelcomeCard } from './WelcomeCard';
+export { default as CurrenciesList } from './CurrenciesList';
+export { default as CurrencyCard } from './CurrencyCard';
